refactor: delete UseGameInfo class
